refactor: Refactoring the code
