+++ nuestra_senora_informativa/src/Services/ServiceMemberSection.ts
@@ -1,0 +1,8 @@
+import axios from "axios";
+
+const URLBase =  'http://localhost:5074/api';
+
+export const getMemberSection = async () => {
+  const response = await axios.get(`${URLBase}/AssociatesSection`);
+  return response.data;
+};
