+++ nuestra_senora_informativa/src/Types/TypeFormVolunteer.ts
@@ -1,0 +1,11 @@
+export type FormVolunteerCreateDto = {
+    Vn_Name: string;
+Vn_Lastname1 : string;
+ Vn_Lastname2 : string;
+Vn_Cedula : string;
+ Vn_Phone: string;
+Vn_Email : string;
+Delivery_Date :Date;
+  End_Date:Date;
+VoluntarieTypeId: number;
+};
