+++ nuestra_senora_informativa/src/Types/TypesAboutUs.ts
@@ -1,0 +1,8 @@
+export type AboutUsSection = {
+    Id_About_Us: number,
+    Subtitle_About_Us: string,
+    MissionTitle_About_Us: string,
+    MissionDescription_About_Us: string,
+    VisionTitle_About_Us: string,
+    VisionDescription_About_Us: string,
+}
