+++ nuestra_senora_informativa/src/Types/TypesConfirmationModal.ts
@@ -1,0 +1,5 @@
+export type ConfirmationModalProps = {
+    isOpen: boolean;
+    onClose: () => void;
+  }
+
