+++ nuestra_senora_informativa/src/Types/TypesContact.ts
@@ -1,0 +1,5 @@
+export type ContactSection = {
+    id_Contact: number,
+    contact_title: string,
+    contact_url: string
+    }
