+++ nuestra_senora_informativa/src/Types/TypesHero.ts
@@ -1,0 +1,8 @@
+export type HeroSection = {
+    
+    Id_Hero: number,
+    Hero_Title: string,
+    Subtitle_Hero: string,
+    Hero_Image_Url: string,
+    HeroText_Button: string,
+    }
