+++ nuestra_senora_informativa/src/Types/TypesNavbar.ts
@@ -1,0 +1,8 @@
+export type NavbarItem = {
+    id_Nav_It: number;         
+    title_Nav: string;         
+    urlNav: string;             
+    order_Item_Nav: number;     
+    isActive: boolean;         
+    children: NavbarItem[];     
+  };
