+++ nuestra_senora_informativa/src/Types/TypesRegistration.ts
@@ -1,0 +1,9 @@
+export type RegistrationSection = {
+    
+    Id_Registration: number,
+    Registration_MoreInfoPrompt: string,
+    RegistrationText_Button: string,
+    Registration_SupportMessage : string,
+    Registration_Image_Url: string,
+    
+    }
