+++ nuestra_senora_informativa/src/Types/TypesServices.ts
@@ -1,0 +1,9 @@
+export type ServiceSection = {
+    
+    Id_ServiceSection: number,
+    Title_Card_SV: string,
+    Image_Card_SV_Url: string,
+    Description_Card_SV: string,
+    SVText_Button: string,
+    
+    }
