+++ nuestra_senora_informativa/src/Types/TypesSiteSettings.ts
@@ -1,0 +1,7 @@
+export type SiteSettings = {
+    
+    Id_SiteSettings: number,
+    SiteTitle: string,
+    Icon_HGA_Url: string,
+    
+    }
