+++ nuestra_senora_informativa/src/Types/TypesTittle.ts
@@ -1,0 +1,7 @@
+export type TittleSection = {
+    
+    Id_TitleSection: number,
+    Title_Text_Section: string,
+    Description_Section: string,
+    
+    }
